backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
GitPython==3.1.40
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
bandit==1.7.5
semgrep==1.45.0
pylint==3.0.3
black==23.12.1
ruff==0.1.11
stripe==7.4.0
APScheduler==3.10.4
